+++ uefi/CHANGELOG.md
@@ -15,6 +15,8 @@
   returns `Option<Event>` instead of `&Event`.
 - `Http::get_mode_data` doesn't consume a parameter anymore and instead return
   an owned value of type `HttpConfigData`
+- Fixed `Serial::read` to properly read only as many bytes as it can and report
+  that `n` to the caller. This includes graceful handling of `Status::TIMEOUT`.
 
 # uefi - v0.36.1 (2025-11-05)